Moved europepmc utils to separate module
