+++ tests/unit_tests/providers/europepmc/utils_test.py
@@ -1,6 +1,6 @@
 from datetime import date
 
-from sciety_labs.providers.europepmc.async_providers import get_first_published_date_within_dates
+from sciety_labs.providers.europepmc.utils import get_first_published_date_within_dates
 
 
 class TestGetFirstPublishedDateWithinDates:
